requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.1
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
email-validator==2.2.0
python-multipart==0.0.9
google-auth==2.35.0
httpx==0.27.2
psycopg[binary]==3.2.3
